+++ src/foreign/mod.rs
@@ -1,0 +1,18 @@
+mod files;
+mod scripts;
+
+use anyhow::{anyhow, Context, Result};
+
+use crate::{JError, Word};
+
+use files::*;
+use scripts::*;
+
+/// https://www.jsoftware.com/help/dictionary/xmain.htm
+pub fn foreign(l: usize, r: usize, _x: Option<&Word>, y: &Word) -> Result<Word> {
+    match (r, l) {
+        (0, k) => f_load_script(k, y),
+        (1, 1) => f_read_file(y).context("reading file"),
+        _ => Err(JError::NonceError).with_context(|| anyhow!("unsupported foreign: {l}!:{r}")),
+    }
+}
