+++ src/lib.rs
@@ -4,6 +4,7 @@ mod ctx;
 mod empty;
 mod error;
 mod eval;
+mod foreign;
 mod modifiers;
 mod number;
 mod scan;
