+++ tests/smoke.toml
@@ -18,16 +18,6 @@ e3,04,03,01,03,581b,401f,2823,
 e3,04,01,,1027'''
 
 [[runs]]
-expr = "# (&.>) < 'foo'"
-output = '''
-┌─┐
-│3│
-└─┘'''
-encoded = '''
-e3,2,01,,28,
-e3,04,01,,03'''
-
-[[runs]]
 expr = '''#&.> ('foo'; 'ba')'''
 output = '''
 ┌─┬─┐
